Refactor app pages into modular sections
